test: fix OperationReviewForm test after rebase
